Rework des commandes de roll et ajout du cooldown
